refactor(library): extract texture_decode submodule from mod.rs
